Improved passed pawn evaluation;
git-svn-id: https://svn.code.sf.net/p/redqueenchess/code/redqueen@494 865e88f3-cc1b-4e5f-86d1-2a7dcd23a5c5

+++ src/bitboard.h
@@ -555,14 +555,14 @@ inline Square extractLSB(Bitboard& bitboard) {
 }
 
 inline const int upperScore(const int value) {
-	return (((value >> 15) & 1) + int16_t((value) >> 16));
+	return ((int16_t(value >> 15) & 1) + int16_t(value >> 16));
 }
 
 inline const int lowerScore(const int value) {
 	return  int16_t(value & 0xFFFF);
 }
 
-inline const int makeScore(const int16_t upperValue, const int16_t lowerValue) {
+inline const int makeScore(const int upperValue, const int lowerValue) {
 	return  lowerValue+(upperValue<<16);
 }
